+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [v7.0.0]
+
+- Fix - Properties not showing for files that are opened outer workspace folder issue fixed
+- Added `Copy` button to copy `Properties` or `Location` based on setting `properties.settings.copyAction`
+- Added `Selection` details. Now we can see selected `lines`, `words`, `array length`, `object size`, `dom nodes length`
+
 ## [v6.0.0]
 
 - Added settings `properties.settings.relativeToRoot` - to show `Location` and `Directory` relative to workspace folder